clean up includes in world/
