+++ src/scripting/bindings/lua_world.cpp
@@ -3,6 +3,8 @@
 #include "core/object.hpp"
 #include "core/player.hpp"
 #include "world/world_manager.hpp"
+#include "world/room.hpp"
+#include "world/zone.hpp"
 
 #define SOL_ALL_SAFETIES_ON 1
 #include <sol/sol.hpp>
