+++ src/world/game_world.hpp
@@ -1,8 +1,6 @@
 #pragma once
 
 #include "../core/ids.hpp"
-#include "../game/login_system.hpp"
-#include "room.hpp"
 
 #include <asio.hpp>
 #include <chrono>
@@ -13,6 +11,8 @@
 
 // Forward declaration
 class CommandSystem;
+class Room;
+class Player;
 
 /**
  * @brief Central world server managing game state and logic
@@ -29,9 +29,9 @@ class GameWorld : public std::enable_shared_from_this<GameWorld> {
     ~GameWorld() = default;
 
     // Character management functions - used by LoginSystem
-    std::shared_ptr<class Player> create_character(const std::string &name);
-    bool save_character(std::shared_ptr<class Player> player);
-    std::shared_ptr<class Player> load_character(const std::string &name);
+    std::shared_ptr<Player> create_character(const std::string &name);
+    bool save_character(std::shared_ptr<Player> player);
+    std::shared_ptr<Player> load_character(const std::string &name);
 
     // Command system access
     std::shared_ptr<CommandSystem> get_command_system() const;
@@ -41,9 +41,9 @@ class GameWorld : public std::enable_shared_from_this<GameWorld> {
     const std::unordered_map<EntityId, std::shared_ptr<Room>> &get_all_rooms() const { return rooms_; }
 
     // Player management
-    void add_player(std::shared_ptr<class Player> player);
-    void remove_player(std::shared_ptr<class Player> player);
-    std::vector<std::shared_ptr<class Player>> get_online_players() const;
+    void add_player(std::shared_ptr<Player> player);
+    void remove_player(std::shared_ptr<Player> player);
+    std::vector<std::shared_ptr<Player>> get_online_players() const;
 
     // World initialization and lifecycle
     void initialize_world();
@@ -65,7 +65,7 @@ class GameWorld : public std::enable_shared_from_this<GameWorld> {
 
     // Core systems
     mutable std::shared_ptr<CommandSystem> command_system_;
-    
+
     // Asio context and timer for game loop
     asio::io_context& io_context_;
     std::unique_ptr<asio::steady_timer> update_timer_;
@@ -74,7 +74,7 @@ class GameWorld : public std::enable_shared_from_this<GameWorld> {
 
     // World data
     std::unordered_map<EntityId, std::shared_ptr<Room>> rooms_;
-    std::vector<std::shared_ptr<class Player>> online_players_;
+    std::vector<std::shared_ptr<Player>> online_players_;
 
     // World state
     bool world_loaded_ = false;
