+++ src/world/room.hpp
@@ -4,6 +4,7 @@
 #include "core/result.hpp"
 #include "core/ids.hpp"
 #include "database/generated/db_room.hpp"
+#include "flags.hpp"
 
 #include <memory>
 #include <vector>
@@ -34,36 +35,6 @@ class Zone;
  * This ensures the game and database stay in sync automatically.
  */
 using Direction = db::Direction;
-
-/** Room sector types affecting movement and behavior */
-enum class SectorType {
-    Inside = 0,
-    City,
-    Field,
-    Forest,
-    Hills,
-    Mountains,
-    Water_Swim,
-    Water_Noswim,
-    Underwater,
-    Flying,
-    Desert,
-    Swamp,
-    Beach,
-    Road,
-    Underground,
-    Lava,
-    Ice,
-    Astral,
-    Fire,
-    Lightning,
-    Spirit,
-    Badlands,
-    Void,
-
-    // Sentinel
-    Undefined
-};
 
 // RoomFlag enum REMOVED - replaced by:
 // - base_light_level_ field for lighting (numeric, more flexible)
