+++ src/world/time_system.cpp
@@ -1,5 +1,6 @@
 #include "time_system.hpp"
-#include "../core/logging.hpp"
+#include "weather.hpp"
+#include "core/logging.hpp"
 #include <fmt/format.h>
 
 using namespace TimeConstants;
